+++ src/koubou/__init__.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     # Python < 3.8 fallback
     try:
-        from importlib_metadata import version
+        from importlib_metadata import version  # type: ignore[no-redef]
 
         __version__ = version("koubou")
     except ImportError: